+++ packages/browser/src/integrations/index.ts
@@ -3,7 +3,3 @@ export { TryCatch } from './trycatch';
 export { Breadcrumbs } from './breadcrumbs';
 export { LinkedErrors } from './linkederrors';
 export { UserAgent } from './useragent';
-
-export { Ember } from './pluggable/ember';
-export { Vue } from './pluggable/vue';
-export { ReportingObserver } from './pluggable/reportingobserver';