+++ packages/core/src/integrations/index.ts
@@ -2,6 +2,3 @@ export { Dedupe } from './dedupe';
 export { FunctionToString } from './functiontostring';
 export { InboundFilters } from './inboundfilters';
 export { ExtraErrorData } from './extraerrordata';
-
-export { Debug } from './pluggable/debug';
-export { RewriteFrames } from './pluggable/rewriteframes';