+++ packages/node/src/integrations/index.ts
@@ -3,6 +3,3 @@ export { Http } from './http';
 export { OnUncaughtException } from './onuncaughtexception';
 export { OnUnhandledRejection } from './onunhandledrejection';
 export { LinkedErrors } from './linkederrors';
-
-export { Modules } from './pluggable/modules';
-export { Transaction } from './pluggable/transaction';